Fetch event rows only for tables being replicated.
pymysqlreplication does not support MariaDB-exclusive binary formats,
e.g. the high-res datetime format used before 10.1.2. In these cases,
attempting to fetch rows from a binlog event will result in a runtime
error.

This commit only attempts to fetch rows *after* determining whether or
not a binlog event should be skipped, so it allows compatibility with
MariaDB databases that have unsupported binary formats so long as they
are not in the tables being replicated.