Fix ci_string parser: convert argument string upper-case before comparison
Previously, when a string with a lowercase letter was passed as the
argument, it won't match any string because it is left unchanged in
comparison. It still worked when the input string was all uppercase,
and that behaviour is unchanged. This patch fixes the case when
a string with a lowercase letter is passed as the argument, and converts
it to uppercase before doing a string match. The returned value in case
of a match is unchanged and is not made uppercase.